move network listener to plugin and rename to offline
