+++ core/src/main/java/com/amplitude/core/platform/EventPipeline.kt
@@ -100,7 +100,7 @@ class EventPipeline(
             }
 
             // Skip flush when offline
-            if (!amplitude.configuration.isNetworkConnected) {
+            if (amplitude.configuration.offline) {
                 continue
             }
